Replace old main activity
Finalize merges
