+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity2"/>
+        <activity android:name=".MainActivity"/>
 
         <meta-data
             android:name="com.facebook.sdk.ApplicationId"
